improvements to user group ui
