+++ frontend/src/views/Org/UserPage/components/UserProjectsSection/UserGroupsSection.tsx
@@ -6,6 +6,7 @@ import { useRemoveUserFromGroup } from "@app/hooks/api";
 import { OrgUser } from "@app/hooks/api/users/types";
 import { usePopUp } from "@app/hooks/usePopUp";
 
+import { UserGroupsProjectsModal } from "./UserGroupProjectsModal";
 import { UserGroupsTable } from "./UserGroupsTable";
 
 type Props = {
@@ -14,7 +15,8 @@ type Props = {
 
 export const UserGroupsSection = ({ orgMembership }: Props) => {
   const { popUp, handlePopUpOpen, handlePopUpClose, handlePopUpToggle } = usePopUp([
-    "removeUserFromGroup"
+    "removeUserFromGroup",
+    "userGroupsProjects"
   ] as const);
 
   const { mutateAsync: removeUserFromGroup } = useRemoveUserFromGroup();
@@ -63,6 +65,8 @@ export const UserGroupsSection = ({ orgMembership }: Props) => {
           return handleRemoveUserFromGroup(popupData.groupSlug);
         }}
       />
+
+      <UserGroupsProjectsModal handlePopUpToggle={handlePopUpToggle} popUp={popUp} />
     </>
   );
 };
